Move BlockArray to scico.numpy, like DeviceArray in jax.numpy
