+++ scico/linop/optics.py
@@ -60,8 +60,8 @@ from numpy.lib.scimath import sqrt  # complex sqrt
 import jax
 
 import scico.numpy as snp
-from scico.array import no_nan_divide
 from scico.linop import Diagonal, Identity, LinearOperator
+from scico.numpy import no_nan_divide
 from scico.typing import Shape
 
 from ._dft import DFT
