+++ scico/operator/biconvolve.py
@@ -13,9 +13,8 @@ import numpy as np
 from jax.scipy.signal import convolve
 
 from scico._generic_operators import LinearOperator, Operator
-from scico.array import is_nested
-from scico.blockarray import BlockArray
 from scico.linop import Convolve, ConvolveByX
+from scico.numpy import BlockArray, is_nested
 from scico.typing import BlockShape, DType, JaxArray
