+++ scico/optimize/_primaldual.py
@@ -15,10 +15,10 @@ from typing import Callable, Optional, Union
 
 import scico.numpy as snp
 from scico.array import ensure_on_device
-from scico.blockarray import BlockArray
 from scico.diagnostics import IterationStats
 from scico.functional import Functional
 from scico.linop import LinearOperator
+from scico.numpy import BlockArray
 from scico.numpy.linalg import norm
 from scico.typing import JaxArray
 from scico.util import Timer
