+++ scico/optimize/pgm.py
@@ -17,10 +17,10 @@ import jax
 
 import scico.numpy as snp
 from scico.array import ensure_on_device
-from scico.blockarray import BlockArray
 from scico.diagnostics import IterationStats
 from scico.functional import Functional
 from scico.loss import Loss
+from scico.numpy import BlockArray
 from scico.typing import JaxArray
 from scico.util import Timer
