+++ scico/test/functional/test_loss.py
@@ -7,12 +7,10 @@ import pytest
 # enable 64-bit mode for output dtype checks
 config.update("jax_enable_x64", True)
 
-
 from prox import prox_test
-
 import scico.numpy as snp
 from scico import functional, linop, loss
-from scico.array import complex_dtype
+from scico.numpy import complex_dtype
 from scico.random import randn, uniform
