+++ scico/test/test_array.py
@@ -7,19 +7,16 @@ from jax.interpreters.xla import DeviceArray
 import pytest
 
 import scico.numpy as snp
-from scico.array import (
+from scico.array import ensure_on_device, indexed_shape, parse_axes, slice_length
+from scico.numpy import (
+    BlockArray,
     complex_dtype,
-    ensure_on_device,
-    indexed_shape,
     is_complex_dtype,
     is_nested,
     is_real_dtype,
     no_nan_divide,
-    parse_axes,
     real_dtype,
-    slice_length,
 )
-from scico.blockarray import BlockArray
 from scico.random import randn
